+++ R/shiny.R
@@ -47,14 +47,14 @@ DTOutput = dataTableOutput
 #'   \code{FALSE}, then the entire data frame is sent to the browser at once.
 #'   Highly recommended for medium to large data frames, which can cause
 #'   browsers to slow down or crash.
-#' @param filterFunc (for expert use only) passed to the \code{filter} argument
+#' @param funcFilter (for expert use only) passed to the \code{filter} argument
 #'   of \code{\link{dataTableAjax}()}
 #' @param ... ignored when \code{expr} returns a table widget, and passed as
 #'   additional arguments to \code{datatable()} when \code{expr} returns a data
 #'   object
 renderDataTable = function(
     expr, server = TRUE, env = parent.frame(), quoted = FALSE,
-    filterFunc = dataTablesFilter, ...
+    funcFilter = dataTablesFilter, ...
   ) {
   if (!quoted) expr = substitute(expr)
 
@@ -112,7 +112,7 @@ renderDataTable = function(
       }
 
       if (is.null(options[['ajax']][['url']])) {
-        url = sessionDataURL(outputInfoEnv[["session"]], origData, outputInfoEnv[["outputName"]], filterFunc)
+        url = sessionDataURL(outputInfoEnv[["session"]], origData, outputInfoEnv[["outputName"]], funcFilter)
         options$ajax$url = url
       }
       instance$x$options = fixServerOptions(options)